+++ .vscode/settings.json
@@ -6,7 +6,7 @@
   },
   "autopep8.args": ["--max-line-length=127"],
   "python.analysis.autoImportCompletions": true,
-  "python.analysis.typeCheckingMode": "basic",
+  // "python.analysis.typeCheckingMode": "basic",
   "python.analysis.inlayHints.variableTypes": true,
   "python.analysis.inlayHints.functionReturnTypes": true,
   "python.analysis.inlayHints.pytestParameters": true
