+++ server/services/canvas/stub.py
